% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{convert_ntf}
\alias{convert_ntf}
\title{convert recursively NULL, TRUE, FALSE to Python equivalents None,
True, False}
\usage{
convert_ntf(x)
}
\arguments{
\item{params}{A parameter-argument list as returned by \code{\link[=get_args_man]{get_args_man()}} or
\code{\link[=pass_args]{pass_args()}}.}
}
\description{
convert recursively NULL, TRUE, FALSE to Python equivalents None,
True, False
}
\examples{
\dontrun{
library("RQGIS3")
params = get_args_man("native:centroids")
convert_ntf(params)
# and just to show that it also works recursively
params$INPUT = list("None", "None")
convert_ntf(params)
}
}
\author{
Jannes Muenchow
}
\keyword{internal}
